requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-openai>=0.0.5
chromadb>=0.4.22
pinecone>=3.0.0
weaviate-client>=3.24.2
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
openai>=1.12.0
tiktoken>=0.5.2
unstructured>=0.11.0
aiofiles<24.0,>=22.0
protobuf<5,>=3.20
boto3>=1.34.0 